remove ink
