--- packages/reporters/cli/src/index.js
@@ -1,8 +1,0 @@
-// @flow strict-local
-
-// $FlowFixMe This is fine.
-const isTTY = process.stdout.isTTY;
-
-export default isTTY
-  ? require('./CLIReporter').default
-  : require('./SimpleCLIReporter').default;
